Refactor & cleanup of registrar code splitting mono vs dotnet code into it's own files
