+++ SimpleConfigSections/ConfigurationElementCollectionForInterface.cs
@@ -11,6 +11,7 @@ namespace SimpleConfigSections
         private readonly Type _elementType;
         private readonly Type _listType;
         private readonly CacheCallback<int, IList> _list;
+		private readonly ConfigurationElementRegistrar _registrar = ConfigurationElementRegistrar.Instance;
 
 		protected ConfigurationElementCollectionForInterface(Type elementType)
         {
@@ -21,8 +22,6 @@ namespace SimpleConfigSections
                                                                 _elementType
                                                             });
             _list = new CacheCallback<int, IList>(ignored=> CreateElements());
-
-			Init();
         }
 
 
@@ -56,13 +55,14 @@ namespace SimpleConfigSections
 
 		protected override void Reset(ConfigurationElement parentElement)
 		{
+			// XXX: Avoid infinite loop on mono.
 			if (!ReflectionHelpers.RunningOnMono)
 				base.Reset(parentElement);
 		}
 
 		protected override void Init()
 		{
-			ConfigurationElementRegistrar.Register(this, _elementType);
+			_registrar.Register(this, _elementType);
 
 			base.Init();
 		}
